requirements-dev.txt
nose==1.3.7
flask-shell-bpython==1.0.0
